refactor: Extract out reducer tests to reducer.spec.tsx
